+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Dependency updates
   - toml 0.4 -> 0.5
   - vulkano-glyph 0.3.0 -> 0.4.0
+- Logarithmic scale, FFT updates
 
 ## 0.1.0
 ### Added